fix(manifest-comparison): prevent overflow in Comparison Report skipped Bates Number totals (#831)
- Widen TotalSkippedBates property in ComparisonModels.cs from int to long
- Use 64-bit checked scalar arithmetic for single and aggregate gap accumulation in BatesRangeAnalyzer
- Guard sequence and supplemental gap boundary calculations (next - current > 1, minNew - maxPrior > 1) against 64-bit overflow
- Throw descriptive OverflowException when total skipped Bates numbers exceed Int64.MaxValue
- Validate inverted gap ranges (v2 < v1) with ArgumentException
- Apply case-insensitive prefix comparison (StringComparison.OrdinalIgnoreCase) and null-defensive iteration in ResultSummaryBatesAnalysis
- Update REQ-162 and REQ-163 in Requirements.md and document comparison report schema in docs/advanced-guide.md
- Add unit and integration tests for single large gap (> Int32.MaxValue), aggregate overflow across multiple smaller gaps, zero-gap inputs, boundaries, and JSON report serialization

+++ src/ManifestComparison/ComparisonModels.cs
@@ -108,7 +108,7 @@ public class BatesAnalysis
     public List<BatesRangeReport> Overlaps { get; set; } = new();
 
     [JsonPropertyName("totalSkippedBates")]
-    public int TotalSkippedBates { get; set; }
+    public long TotalSkippedBates { get; set; }
 
     [JsonPropertyName("priorRangesByProductionSet")]
     public Dictionary<string, string> PriorRangesByProductionSet { get; set; } = new(StringComparer.OrdinalIgnoreCase);
